+++ src/aux-vm-node/managers/AuxUserAuthorizer.ts
@@ -168,8 +168,23 @@ export class AuxUserAuthorizer implements AuxChannelAuthorizer {
         context: FileCalculationContext,
         file: File
     ): ChannelInfo {
+        let channelId = calculateFileValue(context, file, 'aux.channel');
+
+        if (channelId === undefined) {
+            return {
+                id: undefined,
+                locked: calculateBooleanTagValue(
+                    context,
+                    file,
+                    'aux.channel.locked',
+                    false
+                ),
+                maxUsers: getChannelMaxDevicesAllowed(context, file),
+            };
+        }
+
         return {
-            id: calculateFileValue(context, file, 'aux.channel'),
+            id: calculateFileValue(context, file, 'aux.channel').toString(),
             locked: calculateBooleanTagValue(
                 context,
                 file,