+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # AUX Changelog
 
+## V0.9.29
+
+### Date: 08/16/2019
+
+### Changes:
+
+-   Improvements
+    -   Changed `hasFileInInventory()` function to `hasBotInInventory()`.
+    -   Changed `onMerge()` action tag to `onMod()`.
+    -   Changed `aux._editingFile` hidden tag to `aux._editingBot`.
+    -   Gave the player inventory an offset from the bottom of the window so that it is floating.
+    -   Deselecting one of 2 bots in multiselection mode will return the the sheet to single selection mode.
+    -   Removed the direct aux view for now.
+    -   Added new feature in sheet where clicking on a bot's tag will select all bots with that tag.
+-   Bug Fixes
+    -   Resolved issue of the `getBot()` function not working in the search bar.
+    -   Allow the use of a channelID made up entirely of numbers.
+    -   Resolved issue of `setTag()` not working with multiple files when fed a false or null value to set.
+    -   Deleting a bot when in multiselection mode will no longer close the sheet.
+    -   The `onPointerExit()` function will now execute before an `onPointerEnter()` function when hovering over multiple bots.
+    -   Fixed issue in the `RemoveTags()` function where providing a string with a `.` in its tag section failed to remove the correct tags.
+    -   The tag `aux.context` can now be set to a value type of boolean or number.
+    -   Increased the timeout time on the `Create Channel` toast message to give it more processing time so it works more consistently.
+
 ## V0.9.28
 
 ### Date: 08/16/2019